attending guests bubles functional and added to host event details view, cleaned up host event details view
