+++ Bouncr/Views/Host Views/AttendingGuestsBubbles.swift
@@ -8,7 +8,38 @@
 import SwiftUI
 
 struct AttendingGuestsBubbles: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+  @EnvironmentObject var mainController: MainController
+  @ObservedObject var otherUserController: OtherUserController
+  var event: Event
+  
+  //Constants
+  let rows = [GridItem(.fixed(30))]
+
+  var body: some View {
+    
+    //Header on left:
+    Text("Accepted Guests: ")
+      .foregroundColor(Color(red: 66/255, green: 0, blue: 1.0, opacity: 1.0))
+    
+    LazyHGrid(rows: rows, alignment: .lastTextBaseline) {
+      
+      if otherUserController.acceptedInvitesOtherUserArray.count > 4 {
+        ForEach(0..<3){ index in
+          GuestBubble(user: otherUserController.acceptedInvitesOtherUserArray[index])
+        }//End ForEach
+        
+        SeeMoreButton(event: event)
+      }//End if
+      else {
+        ForEach(0..<otherUserController.acceptedInvitesOtherUserArray.count, id: \.self){ index in
+          GuestBubble(user: otherUserController.acceptedInvitesOtherUserArray[index])
+        }//End ForEach
+      }//End else
+      
+    } //End LazyHGrid
+    .onAppear() {
+      mainController.otherUserController.getAcceptedInvitesGuests(eventID: event.id)
     }
+
+  } //End var body: some View
 }
